+++ features/__init__.py
@@ -1,0 +1,1 @@
+"""Feature modules bundled with Script Toolkit."""